child_process: don't emit same handle twice
It's possible to read multiple messages off the parent/child channel.
When that happens, make sure that recvHandle is cleared after emitting
the first message so it doesn't get emitted twice.

It remains to be investigated if this means that it's possible for
handles to either get dropped or end up in the wrong message.